+++ finder/api.py
@@ -1,7 +1,51 @@
 # -*- coding: utf-8 -*-
-# Module used in searching the given pattern in the text provided.
 
 import os
+import threading
+import logging as log
+
+
+log.basicConfig(format='%(message)s', level=log.INFO)
+
+
+class FileReader(threading.Thread):
+    """Thread class for finding the given pattern in a given file.
+
+    Args:
+        path (str): File path.
+        pattern (str): A string to be searched in the given file path.
+
+    .. versionadded:: TODO
+    """
+    def __init__(self, path=None, pattern=None):
+        self.path = path
+        self.pattern = pattern
+        threading.Thread.__init__(self)
+
+    def run(self):
+        """Runs methods for reading the file and searching the pattern in a
+        line of text from the file.
+        """
+        for line_number, line in read(self.path):
+            if search(line, self.pattern):
+                log.info('{path}:{line_number}: {line}'
+                         .format(path=self.path,
+                                 line_number=line_number,
+                                 line=line))
+
+
+def find(paths, pattern=None):
+    """Main method for finding the pattern in the given file paths.
+
+    Args:
+        paths (list): List of paths to be walked through.
+        pattern (str): A string to be searched in the given file path.
+
+    .. versionadded:: TODO
+    """
+    for path in iterfiles(paths):
+        reader = FileReader(path, pattern)
+        reader.start()
 
 
 def search(text=None, pattern=None):
@@ -52,12 +96,18 @@ def iterfiles(*paths):
     .. versionadded:: TODO
     """
     for path in paths:
+        path = os.path.expanduser(path) if not os.path.isabs(path) else path
+
         if os.path.isfile(path) and not is_executable(path):
             yield path
         elif os.path.isdir(path):
             for root, dirs, files in os.walk(path):
                 for file in files:
-                    yield os.path.join(root, file)
+                    file_path = os.path.join(root, file)
+
+                    if (os.path.isfile(file_path) and
+                            not is_executable(file_path)):
+                        yield file_path
 
 
 def is_executable(path):
@@ -71,4 +121,4 @@ def is_executable(path):
 
     .. versionadded:: TODO
     """
-    return os.path.isfile(path) and os.access(path, os.X_OK)
+    return os.access(path, os.X_OK)