+++ CHANGELOG.rst
@@ -6,3 +6,5 @@ Changelog
 - Add ``iterfiles` method to yield all the file paths in a given folder path.
 - Add ``is_executable`` method to validate whether the given file is a executable or not.
 - Add ``read`` method to read a given file line by line.
+- Add wrapper method ``find`` to iterate through the given list of files/directories and find the given pattern in the files.
+- Add ``FileReader`` class to searching all the files concurrently.